Commands: input parses command slug
Selectors can have an optional "slug" parameter now, that can be used to reference that specific selector. If a slug is not found, it's created based on a hash of the entire selector input string.

This is not used yet, but will later be used for data persistence via querystrings.