Command: refactor command input parsing
Now, instead of reading the command and keeping it as a string, we keep it in its own intermediate, AST-ish format around, and pass the needed data as needed, instead of constantly re-parsing the command line (even if cached).

This allows us to keep functionality inside of CommandInput (via a new Class as well) rather than expose a bunch of functions for manipulation.

The encoder is also less aware of our internal format; it simply uses the commands passed, and file information such as names and content.

Closes #30.